### 2023-03-15 #5231 @Bisaloo
Changes since last publication:

```sh
git diff v2.0.0 v3.0.0 -w --ignore-blank-lines --shortstat ':!*/docs_build/*'
```

> 200 files changed, 16070 insertions(+), 8162 deletions(-)

Of which code changes:

```sh
git diff v2.0.0 v3.0.0 -w --ignore-blank-lines --shortstat ':!*/docs_build/*' ':!*/docs/*'
```

> 137 files changed, 15183 insertions(+), 7932 deletions(-)

--

I have a couple of additional questions:

- @Euan-Furness, I see you did not make any direct code contributions to GitHub. Could you clarify your role in the project please?
- Your first publication apparently did not lead to a lot of uses outside your research group. Do you expect the new features of v3 to attract more external users?
