### 2023-03-15 #5231 @Euan-Furness
To your first question: I have designed several of the new features, and wrote their code. Most notably, I designed and coded the ecological interactions and associated logging and GUI features. I'm not sure why that isn't apparent from the commit history; perhaps because it was done in a private development repository, and that code was moved to this public repository by Russell?

To the second: it's true that we haven't been able to attract as much attention as we might have liked from other research groups, although we've been making some progress: we've had expressions of interest from other researchers at conferences and in emails, so we're hopeful that we'll start to see other groups picking this code up and publishing with it in the coming years. Certainly I've been extolling its virtues to other workers in the field. The new features of v3 should help considerably with this: the masks, interactions, and pathogens features each open up considerable new areas of research potential that should make the code appeal to a greater number of outside researchers; particularly ecologists, who may have understandably viewed the previous release as too simplified to ask the questions that they wanted to ask.
