### 2023-03-15 #5231 @Bisaloo
Ok, great, thanks for your answer! I think this submission is a good fit and we can now get properly started.

The next step is now to find at least 2 reviewers. If you have any suggestions for potential reviewers, please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit?usp=sharing) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list). I'll already start looking in my own network in the meantime.
