### 2023-03-16 #5231 @Euan-Furness
I've done some searching: Oskar Hagen (oskar@hagen.bio) would definitely be appropriate as a computational ecologist, although I have no idea how fluent they are with C++.

From the list of reviewers you have provided, there are several who would seem appropriate. The top two, based on research interests, would definitely be bramvandjik88 and emilydolson. After them, I would suggest mtomasini, hvoltbb, slwu89, mhesselbarth, and r-barnes as appropriate as well. All are ecology-inclined, and should be able to review C++ code.
