### 2023-03-16 #5231 @bramvandijk88
Dear Hugo Gruson,

Sure, I can do that. I've read the paper just now and would be more than
happy to walk through the checklist and try out the software. What deadline
are we on?

Best,

Bram

Op do 16 mrt 2023 om 12:01 schreef Hugo Gruson ***@***.***>:

> Hi @bramvandijk88 <https://github.com/bramvandijk88> 👋, would you be
> interested in reviewing this piece of software
> <https://github.com/palaeoware/revosim> and the associated paper
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.05231/joss.05231/10.21105.joss.05231.pdf>
> for the Journal of Open Source Software (JOSS)?
>
> REvoSim v3.0.0: A fast evolutionary simulation tool with ecological
> processes
>
> You have been recommended by one of the authors and I remember your great
> work on the Cacatoo toolbox. As such, I believe you could bring great
> insights for this submission.
>
> You can have a look at our review criteria
> <https://joss.readthedocs.io/en/latest/review_criteria.html> and review
> checklist <https://joss.readthedocs.io/en/latest/review_checklist.html>
> to get an idea of what is expected.
>
> If you are not available to review this submission, do you have any
> recommendation for suitable reviewers?
>
