### 2023-03-20 #5231 @Bisaloo
Hi @emilydolson :wave:, would you be interested in reviewing this [piece of software](https://github.com/palaeoware/revosim) and [the associated paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05231/joss.05231/10.21105.joss.05231.pdf) for the Journal of Open Source Software (JOSS)?

> REvoSim v3.0.0: A fast evolutionary simulation tool with ecological processes

I remember you already reviewed for us last year and I want to be respectful or your time but if you have some availability, I believe you would be a great fit for this submission.
Alternatively, do you have any recommendation for suitable reviewers?
