### 2023-03-20 #5284 @editorialbot
[REVIEW]: REvoSim v3.0.0: A fast evolutionary simulation tool with ecological processes
**Submitting author:** @Euan-Furness (<a href="http://orcid.org/0000-0001-7917-2304">Euan Furness</a>)
**Repository:** https://github.com/palaeoware/revosim
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v3.0.1
