### 2023-03-20 palaeoware/revosim#9 @emilydolson
Compilation error on Linux
Hi, I'm reviewing [the JOSS paper on this package](https://github.com/openjournals/joss-reviews/issues/5284) and I ran into the following compilation error when following the instructions for Linux (Ubuntu >= 21.04) installation. Specifically, this occurred when running `make`:

```
g++ -c -pipe -O2 -std=gnu++11 -Wall -W -D_REENTRANT -fPIC -DSOFTWARE_VERSION=\"3.0.0\" -DQT_NO_DEBUG -DQT_WIDGETS_LIB -DQT_GUI_LIB -DQT_CONCURRENT_LIB -DQT_CORE_LIB -I. -I../../../anaconda3/include/qt -I../../../anaconda3/include/qt/QtWidgets -I../../../anaconda3/include/qt/QtGui -I../../../anaconda3/include/qt/QtConcurrent -I../../../anaconda3/include/qt/QtCore -Ibuild -I. -I../../../anaconda3/mkspecs/linux-g++ -o build/main.o main.cpp
In file included from simmanager.h:35,
                 from genomecomparison.h:21,
                 from mainwindow.h:23,
                 from main.cpp:19:
randoms.h:4:10: fatal error: QRandomGenerator: No such file or directory
    4 | #include <QRandomGenerator>
```

In case it matters, I am running PopOS 22.04.
