### 2023-03-21 palaeoware/revosim#9 @RussellGarwood
Thanks for taking the time to review our paper - we really appreciate your time and input.

This looks like, in this case, the default qmake in your terminal is one in an Anaconda package, and the version of Qt in Anaconda is <5.10 (when QRandomGenerator was introduced). There are, I think, two solutions. Using a fresh clone (or after removing the previous build files from the build directory):

1. Open the .pro in Qt creator (which can be launched from the menu), under tools->options->kits select the version of Qt that was installed in the earlier part of the instructions for building REvoSim, which will be >5.10, then build by hitting ctrl + R

![image](https://user-images.githubusercontent.com/1557073/226686875-6594bf24-6de0-40f5-8d8b-0c99f9f5bb8a.png)

2. Point the terminal to the qmake that was installed by the Qt packages -
2a. Run
`dpkg-query -L qt5-qmake `
to find out where the package placed qmake - in my installation, towards the very bottom of the output of this command I see
`/usr/lib/x86_64-linux-gnu/qt5/bin/qmake`
2b. Cd to the source folder (i.e. that with revosim.pro in it)
2c. Run qmake, specifying the version of qmake from step 2a, in my case
` /usr/lib/x86_64-linux-gnu/qt5/bin/qmake`
2d. Continue following the instructions by running make

I have tested solution 2, which worked, on a machine with Anaconda installed that threw the same error (but which I don't have access to as I am travelling, so can't test number one). If this works, I can add this to the build instructions in the readme. Please do let me know either way!
