### 2023-03-21 palaeoware/revosim#9 @RussellGarwood
Great news! I have added a note to reflect this potential issue to the readme, and linked it to this issue (which I'll now close).
