### 2023-03-27 palaeoware/revosim#10 @bramvandijk88
Installation trouble shooting
> I'm afraid I am stuck at installation. First I attempted Ubuntu, but it appears that either (i) archive.ubuntu is offline, or (ii) this domain is blocked by our institute. I've sent an email to IT, and I'll keep you updated.

Whoops, seems like it's been a while since I last run sudo apt-get update. Problem above is solved. Perhaps mention this is a possible extra step that needs to be taken (sudo apt-get update).

Also, perhaps still address the windows issue below, because I don't think the current explanation is thorough enough. In the documentation it even says "follow standard QT protocol" which many users (me) will not be familiar with.

 I tried the windows protocol, since I also have access to that OS. But having arrived at step 8, I have no clue what to do next. (the explanation of step 8 is quite lacking in details). QT is installed using MYSYS2. Now what?
