### 2023-03-27 #5284 @bramvandijk88
I'm having some issues installing on both Ubuntu and Windows. See https://github.com/palaeoware/revosim/issues/10
