### 2023-03-27 palaeoware/revosim#11 @bramvandijk88
Pathogen layer visualisation does not refresh on "reset"
Expected behaviour: upon hitting reset, the pathogen layer goes back to the initial seed. Currently, it appears to remain in place.
