### 2023-03-27 #5284 @bramvandijk88
Alright I have spent one or two hours with the software, and it seems to be functional. I think found what appears to be a minor bug with the pathogen layer not being reset, but that should be easy to fix. I also tried to assess the other newly implemented functions, but I can't be sure it actually does what the authors claim without spending hours (debugging is a full-time job :))

In any case, I have left two checkboxes unticked.

Firstly, as mentioned in the above comment, I found the installation instruction for Windows lacking. I am not a Windows person, so it could be that developers that like to use Windows (however many people that may be) are familiar with the protocol of using a QT environment with windows. However, in the steps described the final step (8) falls short. I know using C++ based software cross platform can be a pain in the neck, but if you want to have the claim your software can be installed across different operating systems, I would appreciate a little more explanation here. I also ran into an issue installing it on Ubuntu, which honestly was  silly error on my side. That said, why not add the sentence "don't forget to run sudo apt-get update" and prevent the confusion altogether.

Secondly, I left the automated tests box unticked. I am not entirely sure this is necessary for this project, but it feels weird to tick the box if they are not there. I leave this up to the handling editor.

Then a small note to the authors: great job with this software. I have experimented with the earlier version, and the pathogen extension really excites me. I think a lot of biology starts making much more sense in light of pathogens/parasites, and I am looking forward to future publications with this framework. In fact, don't hesitate to suggest me as a reviewer for such future work! If the installation instructions are fixed (Windows step 8 more expanded, Linux reference to sudo apt-get update), I will tick these last boxes and I am supportive of publication.
