### 2023-03-28 8961082
Update readme - Linux instructions
Update Linux build instructions to include command to update package lists
