### 2023-03-28 b564803
Update readme windows instructions
Add further instructions for the final steps of a windows build
