### 2023-03-28 palaeoware/revosim#10 @RussellGarwood
Thanks for these suggestions - and for taking the time to review our paper/software! Good shout on both points, I have:
1. Updated the Linux build instructions to include an update command at the start (though why this was causing the error messages you got is a new one on me)
2. Updating the Windows instructions. I don't currently have a windows machine to hand, so I hope what I have added, based on online documentation and memory covers what is required. I would welcome any further suggested additions, however, if there is anything that is not obvious!
