### 2023-04-28 #5284 @Bisaloo
Hi @Euan-Furness @RussellGarwood :wave: :wave:, I would encourage you to already try to address the comments from the reviewers to keep the ball rolling.
There is no need to wait until reviewers are fully done with their reviews in JOSS, as we aim at having a more conversational and more flexible review process than traditional journals.

> Secondly, I left the automated tests box unticked. I am not entirely sure this is necessary for this project, but it feels weird to tick the box if they are not there. I leave this up to the handling editor.

Thanks, this is a great point! I do agree a full end-to-end testing suite would be very complicated and possibly not a priority for this project. However, I do believe that having continuous integration to verify that the software can be installed on all platforms given your instructions would be a serious improvement. I will open an issue in that sense in the repo.
