### 2023-04-28 #5284 @RussellGarwood
Thanks for the great comments @bramvandijk88. This is just to note that I already improved the windows instructions for build in the master branch - given the software comes with both a windows installer and a portable executable, I'm hopeful any windows users that actually chose to build it will be familiar with C++ development on Windows. As such I hope that is dealt with, but would welcome feedback to the contrary if you disagree! (I also added the sudo apt update to the Linux instructions - had the same though).

Great spot on the pathogen layer, thanks for that - a fix should be simple and is on my to do list.

-- Secondly, I left the automated tests box unticked

Indeed, our tests are not automated - but they are present through the GUI, in case you missed this (we wanted those to be visible to users so they could check REvoSim's functioning before using!)

Thanks to drawing my attention to this process @Bisaloo - I've been away for most of the last month and had not flagged all the recent updates!
