### 2023-05-10 palaeoware/revosim#13 @Bisaloo
Add continuous integration to verify installation instructions
It is not uncommon that installation instructions are incomplete or out of date because software authors forgot they already had some system dependencies installed on their machine.

To ensure that the code compiles and that installation instructions are up-to-date, it would be helpful to add continuous integration (e.g., GitHub Actions) that try and install the software.

It's probably too difficult to go one step further and add end-to-end testing at this point because testing graphical interfaces is generally more complex than testing command line interfaces.

Please let me know if you need additional documentation or guidance on this issue.

This is part of https://github.com/openjournals/joss-reviews/issues/5284.
