### 2023-05-12 palaeoware/revosim#14 @emilydolson
Improvements to documentation
As part of my review on the JOSS paper on REvoSim (https://github.com/openjournals/joss-reviews/issues/5284) I have a couple of suggestions for improvements to the documentation. In general, the documentation of all of the features is very thorough. What I think is missing is a layer between the very specific feature documentation and the very general quickstart guide. In particular, it would be really useful to have is a walk-through of the rationale behind the decisions that you make when setting up an experiment in REvoSim. This could be an example of a previous paper done using the platform (I left the "Example usage" box in the JOSS review form unchecked because there doesn't currently  seem to be anything quite like this).

In a similar vein, I think the documentation would benefit from a little more clarification of how certain concepts in REvoSim map onto more general concepts that users may have encountered elsewhere. The two main concepts that I would have loved more clarification on are:
1) What exactly defines a species? How do new species evolve?
2) How do iterations in this system map onto more generalizable units of time, such as generations. I understand that the exact mapping of iterations to generations depends on the configuration, but a rough ball park would be helpful (or, even better, a display somewhere indicating how many generations have elapsed on average).
