### 2023-05-12 #5284 @emilydolson
# Software/documentation

This seems like a useful piece of software, and all of the functionality worked as intended! I have opened an issue suggesting a few improvements to the documentation (including adding an example, as suggested in the review criteria).

I checked the "testing" checkbox because the procedure for running tests is clearly documented and worked for me. That said, I'm also always in favor of adding automated testing where possible (it could just be a github action that automatically runs the same tests that a user could run locally). From a user perspective, if a package has automated testing, I am substantially more likely to trust it enough to use it in my research.

I haven't checked off the "performance" box yet, because (as mentioned in my comments on the documentation) I feel like I don't have a good perspective yet for how iterations in this system map onto units of time in other systems. There aren't especially strong performance claims made in the paper, but they do say the system is efficient enough to ask questions about large scale macroevolutionary processes, so I'd just like to get a rough sense for how quickly it goes through generations before checking this one off.

# Paper

The paper is well-written and does a good job describing the newly added features. The only area where it is lacking is in the description of the current state of the field. I completely agree that this software fills a worthwhile niche (the combination of a full-featured GUI combined with relatively simple/efficient genomes that can be configured in a variety of ways). However, I was surprised to see the statement that "REvoSim is unusual in that it simulates evolutionary processes that occur within populations rather than just between them (Garwood et al., 2019), and is highly efficient, allowing long time spans and large population sizes on modest hardware." I'm definitely biased by the sub-field I'm in, but my impression is there are actually quite a lot of systems with these properties (although they make various tradeoffs between complexity and efficiency). I'm by no means suggesting that the authors need to refer to all of the following, but here are some examples of other systems that the authors may want to be aware of:
- [Symbulation](https://github.com/anyaevostinar/SymbulationEmp) (has some parallels to the pathogen layer in REvoSim)
- [SLiM](https://messerlab.org/slim/)
- [Avida](https://github.com/devosoft/avida)
- [MABE](https://github.com/Hintzelab/MABE)
- [Aevol](https://www.aevol.fr/)

In case it's helpful, I reviewed a bunch more computational evolution platforms [in this paper](https://www.frontiersin.org/articles/10.3389/fevo.2021.750779/full) and am working on assembling more information on them in [this wiki](https://alife.org/encyclopedia/category/software-platforms/).

Again, I don't think all of that needs to go in the paper, but I think it would be worth adjusting the discussion of the current state of the field to make it clearer to the reader what REvoSim's niche within the broader landscape is. In my mind, it does a fantastic job of occupying the space just left of digital evolution on this continuum (which is explained in more detail in the paper I linked to):

![image](https://github.com/openjournals/joss-reviews/assets/3119922/b92ac9f1-6032-4ac5-a91c-a52b73db555f)

Otherwise I'd say this is all looking very good!
