### 2023-05-12 #5284 @Euan-Furness
Hi @Bisaloo

I know you suggested this before, but just to check, are you happy for us to make changes to the text of the paper on Github at this stage, to make improvements suggested by reviewers? I don't want to accidentally put people in a position of looking at lots of different versions of a document.

If this is ok, I will start making suggested changes to the paper to address the current state of the field.
