### 2023-05-12 #5284 @RussellGarwood
Many thanks @emilydolson for the really useful comments - as per the above, @Euan-Furness will amend the paper as suggested, and I have noted the issue relating to documentaiton on the main repo. I'll be addressing all those issues in one go as soon as workload allows.
