### 2023-05-12 #5284 @Bisaloo
> If this is ok, I will start making suggested changes to the paper to address the current state of the field.

Yes, it's fine, please go ahead :+1:. Feel free to [use editorialbot](https://joss.readthedocs.io/en/latest/editorial_bot.html) after to regenerate the paper and post a link here.
