### 2023-05-16 52e2e8b
Update paper.md
Changes to the Statement of Need based on the review by Emily Dolson. We are pushing the word count already, and so a lengthy discussion of the state of the field would be tricky, but I have framed REvoSim's place in terms of simulation and digital evolution tools, as discussed in Dolson & Ofria (2021).
