+++ JOSS/paper.bib
@@ -22,12 +22,28 @@
 	Year = 2020
 }
 
+@article{batut:2013,
+	doi = {10.1186/1471-2105-14-S15-S11},
+	Author = {{Batut}, B. and {Parsons}, D.~P. and {Fischer}, S. and {Beslon}, G. and {Knibbe}, C.},
+	Journal = {BMC Bioinformatics},
+	Title = {{In silico experimental evolution: a tool to test evolutionary scenarios}},
+	Year = 2013
+}
+
 @article{cabral:2017,
 	doi = {10.1111/ecog.02480},
 	Author = {{Cabral}, J.~S. and {Valente}, L. and {Hartig}, F.},
 	Journal = {Ecography},
 	Title = {{Mechanistic simulation models in macroecology and biogeography: state-of-art and prospects}},
 	Year = 2016
+}
+
+@article{dolson:2021,
+	doi = {10.3389/fevo.2021.750779},
+	Author = {{Dolson}, E. and {Ofria}, C.},
+	Journal = {Frontiers in Ecology and Evolution},
+	Title = {{Digital evolution for ecology research: a review}},
+	Year = 2021
 }
 
 @article{garwood:2019,
