### 2023-05-16 91b5d65
Update paper.md
Typo correction.
