### 2023-05-16 f45884b
Update paper.md
Second typo fix.
