### 2023-06-05 12bdcd5
Reset masks on setup run
