### 2023-06-05 650f088
reset pathogens on setup run
