### 2023-06-05 865375c
Started documentation changes requested for JOSS
