### 2023-06-05 b09e5ff
Add warning messages when saving settings and linkages or environment are loaded
