### 2023-06-06 5ddefce
Update paper.md
Additional example useage
