### 2023-06-06 801a959
Add interactions settings stored in cellMasterSettings to save settings
