+++ revosim/mainwindow.cpp
@@ -4553,6 +4553,12 @@ void MainWindow::loadSettings(QString fileName, bool calledFromCommandLine)
                 environmentModeChanged(settingsFileIn.readElementText().toInt(), true);
             if (settingsFileIn.name() == "pathogenMode")
                 simulationManager->simulationSettings->pathogenMode = settingsFileIn.readElementText().toInt();
+            if (settingsFileIn.name() == "interactions")
+                simulationManager->cellSettingsMaster->interactions = settingsFileIn.readElementText().toInt();
+            if (settingsFileIn.name() == "predationEfficiency")
+                simulationManager->cellSettingsMaster->predationEfficiency = settingsFileIn.readElementText().toInt();
+            if (settingsFileIn.name() == "minDeltaPredatorness")
+                simulationManager->cellSettingsMaster->minDeltaPredatorness = settingsFileIn.readElementText().toInt();
 
             //No Gui options for the remaining settings as yet.
             if (settingsFileIn.name() == "speciesSamples")
@@ -4581,8 +4587,6 @@ void MainWindow::loadSettings(QString fileName, bool calledFromCommandLine)
                 simulationManager->cellSettingsMaster->breedDifference = intToBool(settingsFileIn.readElementText().toInt());
             if (settingsFileIn.name() == "breedSpecies")
                 simulationManager->cellSettingsMaster->breedSpecies = settingsFileIn.readElementText().toInt();
-            if (settingsFileIn.name() == "multiBreedList")
-                simulationManager->cellSettingsMaster->multiBreedList = settingsFileIn.readElementText().toInt();
             if (settingsFileIn.name() == "pathOn")
                 simulationManager->cellSettingsMaster->pathOn = settingsFileIn.readElementText().toInt();
             if (settingsFileIn.name() == "variableMutate")
@@ -4605,14 +4609,16 @@ void MainWindow::loadSettings(QString fileName, bool calledFromCommandLine)
                 simulationManager->simulationSettings->gui = intToBool(settingsFileIn.readElementText().toInt());
             if (settingsFileIn.name() == "environmentInterpolate")
                 simulationManager->simulationSettings->environmentInterpolate = intToBool(settingsFileIn.readElementText().toInt());
-            if (settingsFileIn.name() == "interactblocks")
+            if (settingsFileIn.name() == "interactBlocks")
                 simulationManager->cellSettingsMaster->interactBlocks = intToBool(settingsFileIn.readElementText().toInt());
-            if (settingsFileIn.name() == "multibreedlist")
+            if (settingsFileIn.name() == "multiBreedList")
                 simulationManager->cellSettingsMaster->multiBreedList = intToBool(settingsFileIn.readElementText().toInt());
-            if (settingsFileIn.name() == "interactfitness")
+            if (settingsFileIn.name() == "interactFitness")
                 simulationManager->cellSettingsMaster->interactFitness = intToBool(settingsFileIn.readElementText().toInt());
-            if (settingsFileIn.name() == "interactenergy")
+            if (settingsFileIn.name() == "interactEnergy")
                 simulationManager->cellSettingsMaster->interactEnergy = intToBool(settingsFileIn.readElementText().toInt());
+            if (settingsFileIn.name() == "interactWithinSpecies")
+                simulationManager->cellSettingsMaster->interactWithinSpecies = intToBool(settingsFileIn.readElementText().toInt());
             if (settingsFileIn.name() == "v2log")
             {
[…]
