### 2023-06-06 d26acae
Write logs to settings files
