### 2023-06-07 1daba49
updating documentation for JOSS review
