### 2023-06-08 bf31190
Add breedsuccess to log
