### 2023-06-09 1cbfdd9
Fix pathogens command line switch, add one for pathogens evolve
