### 2023-06-13 c2ec456
Added generation calculation to iteration level log
