+++ revosim/logsimulation.cpp
@@ -532,9 +532,9 @@ QString LogSimulation::processLogTextGeneral(QString text)
 
     //RJG - grid level stats
     if (text.contains("*gridNumberAlive*") || text.contains("*gridBreedEntries*") || text.contains("*gridBreedFails*") || text.contains("*gridMeanFitness*") || text.contains("*gridTrophicHistograms*")
-            || text.contains("*gridBreedSuccess*"))
+            || text.contains("*gridBreedSuccess*") || text.contains("*gridGeneration*"))
     {
-        int gridNumberAlive = 0, gridTotalFitness = 0, gridBreedEntries = 0, gridBreedFails = 0, gridBreedSuccess = 0;
+        int gridNumberAlive = 0, gridTotalFitness = 0, gridBreedEntries = 0, gridBreedFails = 0, gridBreedSuccess = 0, gridGeneration = 0;
         int trophicL01 = 0, trophic0102 = 0, trophic0203 = 0, trophic0304 = 0, trophic0405 = 0, trophic0506 = 0, trophic0607 = 0, trophic0708 = 0, trophic0809 = 0, trophic0910 = 0, trophic1011 = 0,
             trophic1112 = 0, trophic1213 = 0, trophic1314 = 0, trophic1415 = 0, trophic1516 = 0, trophic1617 = 0, trophic1718 = 0, trophic1819 = 0, trophic1920 = 0, trophic2021 = 0, trophic2122 = 0,
             trophic2223 = 0, trophic2324 = 0, trophic2425 = 0, trophic2526 = 0, trophic2627 = 0, trophic2728 = 0, trophic2829 = 0, trophic2930 = 0, trophicG30 = 0;
@@ -546,6 +546,8 @@ QString LogSimulation::processLogTextGeneral(QString text)
                 gridBreedEntries += breedAttempts[i][j];
                 gridBreedSuccess += breedSuccess[i][j];
                 gridBreedFails += breedFails[i][j];
+                gridGeneration += breedGeneration[i][j];
+
                 //----RJG: Manually count number alive thanks to maxUsed descendants
                 for  (int k = 0; k < simulationManager->cellSettingsMaster->slotsPerSquare; k++)
                     if (critters[i][j][k].age > 0)
@@ -588,6 +590,9 @@ QString LogSimulation::processLogTextGeneral(QString text)
         text.replace("*gridNumberAlive*", QString::number(gridNumberAlive));
         text.replace("*gridBreedEntries*", QString::number(gridBreedEntries));
         text.replace("*gridBreedSuccess*", QString::number(gridBreedSuccess));
+        //Note the below assumes one breed per generation, and user should check this. Is in docs.
+        if (gridBreedSuccess > 0) text.replace("*gridGeneration*", QString::number(static_cast<double>(gridGeneration) / static_cast<double>(gridBreedSuccess)));
+        else text.replace("*gridGeneration*", QString("Breed success is zero - can't calculate this."));
         text.replace("*gridBreedFails*", QString::number(gridBreedFails));
         text.replace("*gridMeanFitness*", QString::number(gridMeanFitness));
         text.replace("*gridTrophicHistograms*", (QString::number(trophicL01) + "," + QString::number(trophic0102) + "," + QString::number(trophic0203) + "," + QString::number(
