### 2023-06-13 3656987
Fix typos and add stuff to command line, also put new logging tags into logging page
