+++ revosim/docs/commandline.rst
@@ -7,7 +7,7 @@ REvoSim provides command line options to enable full control of simulations from
 
 This document does not provide instructions on how to write a batch file, or determine correct paths (which will differ between operating systems).
 
-Command line switches are documented below, and examples are given of their usage. Any combination of switches in any order is allowed. Note that on some systems, path names with spaces may cause problems; these are best avoided, but alternatively paths can be encased in double quotes if they contain a space.
+Command line switches are documented below, and examples are given of their usage. Any combination of switches in any order is allowed. Note that on some systems, path names with spaces may cause problems: these are best avoided, but alternatively paths can be encased in double quotes if they contain a space.
 
 Note in particular the --auto command, which automatically starts a simulation, and exits the program at its end. This will be required in most batch scenarios, as will the -k switch to turn on one or both types of logging.
 
@@ -19,10 +19,12 @@ On Linux this command (assuming you are in the same folder as the REvoSim binary
 
 ./revosim -e /home/user/environment/ -t=On -k=Normal --auto=50000
 
+The easiest way of modifying multiple variables for runs - and of setting custom log text, which was added in REvoSim v3.0.0 - is to use the software to output a settings files, and then to load this from the command line (--settings).
+
 Running via SSH
 ---------------
 
-In many instances it may be beneficial to run REvoSim on a remote machine. With remote desktopping solutions that provide a GUI, this can be achieved very easily following the normal operating procedures. However, due to the development of REvoSim, launching the software via the command linelaunches the GUI. This means that it is not possible to launch the command line version via SSH unless X-forwarding or another such solution is implemented. However, if no interaction with the GUI is required, there are a number of solutions that can bypass this requirement. The simplest, that will work on the mjaority of systems with Qt installed, is to include the flag -platform offscreen:
+In many instances it may be beneficial to run REvoSim on a remote machine. With remote desktopping solutions that provide a GUI, this can be achieved very easily following the normal operating procedures. However, due to the development of REvoSim, launching the software via the command line launches the GUI. This means that it is not possible to launch the command line version via SSH unless X-forwarding or another such solution is implemented. However, if no interaction with the GUI is required, there are a number of solutions that can bypass this requirement. The simplest, that will work on the majority of systems with Qt installed, is to include the flag -platform offscreen:
 
 ./revosim --auto 1000 -platform offscreen
 
@@ -30,10 +32,11 @@ This co-opts a Qt platform plugin to simply render to an offscreen buffer. Anoth
 
 xvfb-run -a  ./revosim --auto 1000 
 
-You can then use e.g. SFTP to collect output files as required. 
+In any case, once a run is complete, you can then use e.g. SFTP to collect output files as required. 
 
 Single-letter switches
 ----------------------
+
 These can be used as either in single letter form (-a) or long form (\--startage - note the double -). All require a value, which can be separated from the switch using a space or an equals - or in single character mode, without a separator at all (this doesn't work in -- long form, where a separator IS required). Where no parameter is noted in square brackets this is a boolean option, which is either On or Off (you can also use 1/0, true/false, yes/no, y/n - which are not case sensitive). e.g.
 
 - -a30
@@ -48,7 +51,7 @@ For a boolean:
 - -q 1
 - -q=truE
 - -q Y
-- \--recalcfitness False
+- \--recalcfitness True
 - \--recalcfitness = Y
 
 Full list of single-letter switches (also available using switch -h, help):
@@ -66,7 +69,7 @@ Full list of single-letter switches (also available using switch -h, help):
 - -l, --excludenodescendents <On/Off>                     Exclude species without descendents from phylogeny logs
[…]
