+++ revosim/docs/loggingsim.rst
@@ -55,7 +55,7 @@ Iteration keywords
 :\*gridBreedSuccess\*: The number of successful breeds in the polling iteration for all organisms in the grid.
 :\*gridMeanFitness\*: The mean fitness of all the organisms in the grid at polling iteration.
 :\*gridNumberAlive\*: The number of organisms alive at polling iteration.
-:\*gridGeneration\*: This calculates the average age (thus generation time) of all successful parents in the polling iteration. Note that it calculates this based on the assumption, which is true for the majority of settings, that organisms will only successfully breed once in their lifetime. This assumption can be checked using the gridNumberAlive and gridBreedSuccess outputs to calculate the mean number of breeding individuals at any generation.
+:\*gridGeneration\*: This calculates the average age (thus generation time) of all successful parents in the polling iteration. Note that it calculates this based on the assumption (which is true for the majority of settings), that organisms will only successfully breed once in their lifetime. This assumption can be checked using the \*gridNumberAlive\* and \*gridBreedSuccess\* outputs to calculate the mean number of breeding individuals at any generation.
 :\*gridTrophicHistograms\*: This outputs a histogram of the trophic levels of all the organisms in the grid, at 0.1 intervals between 0 and 3.
 :\*iteration\*: The current iteration.
 :\*printSettings\*: This prints a string of all REvoSim settings at the polling iteration, or start of the run if placed in the header.
