### 2023-07-04 palaeoware/revosim#14 @RussellGarwood
Thanks for these suggestions - they have helped us improve the documentation a great deal (the need for such things is harder to spot when you've written the code).

With above commits there is a new section in the REvoSim docs ["3. Concepts and Example Usage"](https://revosim.readthedocs.io/en/latest/). This includes, in terms of concetps:

1. Introduction to the REvoSim species concept
2. A discussion of how to calculate generation time for any given settings, alongside a few additions to the logging code to make this easier for the user to do

In terms of Example usage, there is now a discussion of the thought process behind the experiments report in Furness et al. (2021).

Hopefully this addresses both the above requests - I'll leave the issue open for a bit in case there is anything else that would be useful to add, and then close it in a week or so.
