### 2023-07-06 #5284 @RussellGarwood
Hi all - thanks ever so much for taking the time to review this contribution. With an exception outlined below, we believe that all points have now been addressed:

https://github.com/palaeoware/revosim/issues/11 - Pathogen layer now resets as expected
https://github.com/palaeoware/revosim/issues/10 - More details have been added to the readme for both Linux and windows builds
https://github.com/palaeoware/revosim/issues/9 - Details on how to overcome this Linux environment issue are also now included in the readme
https://github.com/palaeoware/revosim/issues/14 - We have now added details on time, species concepts, and exemplar usage to the documentation

For all of these, more details are included in the closed issues. I note also that @Euan-Furness may wish to highlight the changes he has made to the manuscript. For clarification of a couple of other points from the reviews:

@bramvandijk88 "Firstly, as mentioned in the above comment, I found the installation instruction for Windows lacking. I am not a Windows person, so it could be that developers that like to use Windows (however many people that may be) are familiar with the protocol of using a QT environment with windows." - we just wanted to highlight that we provide both a portable binary, and an installation file for Windows. We have improved the build instructions for that OS, as requested (see comments on issue above), but also expect that any users that want to just use (rather than modify) the software, will not need to build it on windows.

@bramvandijk88 "Secondly, I left the automated tests box unticked. I am not entirely sure this is necessary for this project, but it feels weird to tick the box if they are not there. I leave this up to the handling editor." - Thanks for this - we just wanted to highlight that there are test, although they are currently not automated. We chose to make them visible to the user on the assumption that many of our users will not want to build the package to check the tests. We have made note of this and @emilydolson's testing comment, and will look into automating these tests in future releases as well as displaying the outputs to the user.

@emilydolson "I haven't checked off the "performance" box yet, because (as mentioned in my comments on the documentation) I feel like I don't have a good perspective yet for how iterations in this system map onto units of time in other systems" - just to highlight, that as per [the updated docs](https://revosim.readthedocs.io/en/latest/revosimconcepts.html), you can now check this for any settings you like with ease by using the *gridGeneration* tag in the logging system. Given you might not have time to rebuild RevoSim to try this out, though, here are some stats from my work system (Intel Xeon(R) Gold 6338N CPU @ 2.20GHz × 64). At equilibrium, with default settings, REvoSim has a carrying capacity of ~200,000 individuals, and mean generation time is 11.4 iterations. In such runs, REvoSim is getting through 171.5k iterations per hour, and so my system can manage about ~15000 generations per hour.

@Bisaloo We not - and are grateful for - the suggestion we add continuous integration for REvoSim. We have not - as yet - implemented this for the package. Based on our research, continuous integration for Mac, Windows and Linux for this package would be a time consuming to set up, and then require ongoing maintenance. In this light it's not clear to us the benefits outweight the costs. Our consierations include:
- 1. There are a limited number of people contributing the the code base (currently three of us) and haven't had any issues stemming from a lack of continuous integration to date (see below for comment regarding the build issues our reviewers experienced).
- 2. Given the infrequency of releases for this package (v2.0.0 2019, v3.0.0 2023 - such is the lot of full time academics with limited coding time!) manually testing our installers/binaries for release, which we did on vanilla virtual machines, isn't a big deal!
- 3. The build issues in this review - as an example - were based on the individual Linux environment of the reviewers or a lack of detail in our readme. We have fixed the latter. We provided fixes for the former, and in both cases, continuous integration would not have helped avoid the Linux issues (neither related to missing dependencies).
We do intend to leave the issue open, however, and revisit it to check if the benefits do outweight the time investment  (e.g. if the team expands, or our release schedule increases in frequency). We hope that this is OK.
