### 2023-07-06 palaeoware/revosim#13 @RussellGarwood
This point is appreciated - and we will keep this issue open to reasses at future releases whether we want to set this up, as per response in JOSS review:

https://github.com/openjournals/joss-reviews/issues/5284#issuecomment-1623992476
