### 2023-07-06 3e6b520
Added automated header text for new log tags
