+++ utility/envirogen/ui_mainwindow.h
@@ -1,0 +1,1011 @@
+/********************************************************************************
+** Form generated from reading UI file 'mainwindow.ui'
+**
+** Created by: Qt User Interface Compiler version 5.15.3
+**
+** WARNING! All changes made in this file will be lost when recompiling UI file!
+********************************************************************************/
+
+#ifndef UI_MAINWINDOW_H
+#define UI_MAINWINDOW_H
+
+#include <QtCore/QVariant>
+#include <QtWidgets/QApplication>
+#include <QtWidgets/QCheckBox>
+#include <QtWidgets/QComboBox>
+#include <QtWidgets/QDockWidget>
+#include <QtWidgets/QDoubleSpinBox>
+#include <QtWidgets/QFrame>
+#include <QtWidgets/QGraphicsView>
+#include <QtWidgets/QGridLayout>
+#include <QtWidgets/QLabel>
+#include <QtWidgets/QMainWindow>
+#include <QtWidgets/QMenuBar>
+#include <QtWidgets/QPushButton>
+#include <QtWidgets/QSpacerItem>
+#include <QtWidgets/QSpinBox>
+#include <QtWidgets/QStatusBar>
+#include <QtWidgets/QTabWidget>
+#include <QtWidgets/QTextBrowser>
+#include <QtWidgets/QToolBar>
+#include <QtWidgets/QVBoxLayout>
+#include <QtWidgets/QWidget>
+
+QT_BEGIN_NAMESPACE
+
+class Ui_MainWindow
+{
+public:
+    QWidget *centralWidget;
+    QGridLayout *gridLayout_3;
+    QGraphicsView *enviroView;
+    QMenuBar *menuBar;
+    QToolBar *toolBar;
+    QStatusBar *statusBar;
+    QDockWidget *envSettings;
+    QWidget *dockWidgetContents;
+    QVBoxLayout *verticalLayout;
+    QFrame *dockWidgetFrame;
+    QVBoxLayout *verticalLayout_2;
+    QTabWidget *settings_tab_widget;
+    QWidget *output_tab;
+    QVBoxLayout *verticalLayout_4;
+    QLabel *instructions_label;
+    QLabel *output_label;
+    QCheckBox *save_images_checkbox;
+    QLabel *label_7;
+    QTextBrowser *path;
+    QPushButton *change_path;
+    QLabel *label_5;
+    QSpinBox *numGenerations;
+    QLabel *label_26;
+    QSpinBox *spinSize;
+    QLabel *label_32;
+    QComboBox *environment_comboBox;
+    QLabel *display_label;
+    QCheckBox *resize;
+    QSpacerItem *verticalSpacer;
+    QWidget *russell_tab;
+    QGridLayout *gridLayout_4;
+    QSpinBox *buffSpin;
+    QLabel *russell_label;
+    QLabel *label;
+    QLabel *label_6;
+    QSpinBox *sMin;
+    QDoubleSpinBox *velMax2;
+    QLabel *label_13;
+    QSpinBox *cvelMax;
+    QDoubleSpinBox *convergeAt;
+    QLabel *label_4;
+    QCheckBox *periodic;
+    QLabel *label_12;
+    QSpinBox *numSeed;
+    QSpacerItem *verticalSpacer_6;
+    QLabel *label_3;
+    QLabel *label_11;
+    QDoubleSpinBox *factorSpin;
+    QSpinBox *sMax;
+    QCheckBox *blurChk;
+    QLabel *label_8;
+    QLabel *label_2;
+    QLabel *label_9;
+    QWidget *mark_tab;
+    QGridLayout *gridLayout;
+    QDoubleSpinBox *ME_MaxAccel;
[…]
