### 2023-07-10 #5284 @Euan-Furness
Hi all - as Russell has already said, thank you to all involved in the review process thusfar. Russell's comments on changes were pretty comprehensive, but as a brief addition:

We modified the Statement of Need to incorporate discussion of REvoSim's place within the broader simulation ecosystem, with reference to Dolson & Ofria (2021). We also discussed this in the new addition to the documentation files, which provides an example usage of REvoSim based on an existing paper, and guides the user through the steps necessary to set up and run simulations with a specific goal in mind, generating log files that can be processed for useful information.
