### 2023-07-10 #5284 @Bisaloo
Thanks for your answers @RussellGarwood & @Euan-Furness !

I would still recommend adding CI at some point but I don't think this is a blocker for publication. Indeed, CI would only check that the software compiles and can be installed correctly but would not allow testing of the software functionality without a complex setup.

@bramvandijk88, could you double check that you're happy with the answers and check the remaining boxes in your checklist if so? Thanks!
