### 2023-07-18 108dbc5
Trying to get to bottom of what is going on in analyser - minor changes to formatting probably
