+++ revosim/docs_build/_sources/commandline.rst.txt
@@ -7,33 +7,44 @@ REvoSim provides command line options to enable full control of simulations from
 
 This document does not provide instructions on how to write a batch file, or determine correct paths (which will differ between operating systems).
 
-Command line switches are documented below, and examples are given of their usage. Any combination of switches in any order is allowed. Note that on some systems, path names with spaces may cause problems; these are best avoided, but alternatively paths can be encased in double quotes if they contain a space.
+Command line switches are documented below, and examples are given of their usage. Any combination of switches in any order is allowed. Note that on some systems, path names with spaces may cause problems: these are best avoided, but alternatively paths can be encased in double quotes if they contain a space.
 
 Note in particular the --auto command, which automatically starts a simulation, and exits the program at its end. This will be required in most batch scenarios, as will the -k switch to turn on one or both types of logging.
 
 e.g. a command line to run a similation on the environment files in "c:\\revosim\\data\\env1" for 50000 iterations, using normal (running) logs, and with settings at defaults except for 'toroidal' (on):
 
-c:\\revosim\\revosim.exe -e c:\\revosim\\data\\env1 -t=On -k=Normal --auto=50000
+.. code-block:: console
+    
+    c:\\revosim\\revosim.exe -e c:\\revosim\\data\\env1 -t=On -k=Normal \--auto=50000
 
 On Linux this command (assuming you are in the same folder as the REvoSim binary):
 
-./revosim -e /home/user/environment/ -t=On -k=Normal --auto=50000
+.. code-block:: console
+    
+    ./revosim -e /home/user/environment/ -t=On -k=Normal \--auto=50000
+
+The easiest way of modifying multiple variables for runs - and of setting custom log text, which was added in REvoSim v3.0.0 - is to use the software to output a settings files, and then to load this from the command line (\--settings).
 
 Running via SSH
 ---------------
 
-In many instances it may be beneficial to run REvoSim on a remote machine. With remote desktopping solutions that provide a GUI, this can be achieved very easily following the normal operating procedures. However, due to the development of REvoSim, launching the software via the command linelaunches the GUI. This means that it is not possible to launch the command line version via SSH unless X-forwarding or another such solution is implemented. However, if no interaction with the GUI is required, there are a number of solutions that can bypass this requirement. The simplest, that will work on the mjaority of systems with Qt installed, is to include the flag -platform offscreen:
+In many instances it may be beneficial to run REvoSim on a remote machine. With remote desktopping solutions that provide a GUI, this can be achieved very easily following the normal operating procedures. However, due to the development of REvoSim, launching the software via the command line launches the GUI. This means that it is not possible to launch the command line version via SSH unless X-forwarding or another such solution is implemented. However, if no interaction with the GUI is required, there are a number of solutions that can bypass this requirement. The simplest, that will work on the majority of systems with Qt installed, is to include the flag -platform offscreen:
 
-./revosim --auto 1000 -platform offscreen
+.. code-block:: console
+    
+    /revosim \--auto 1000 -platform offscreen
 
 This co-opts a Qt platform plugin to simply render to an offscreen buffer. Another option is to install XVFB - this is an "X server that can run on machines with no display hardware and no physical input device" (`More information <https://www.x.org/releases/X11R7.6/doc/man/man1/Xvfb.1.xhtml>`_). Once installed, you can launch revosim as follows:
 
-xvfb-run -a  ./revosim --auto 1000 
+.. code-block:: console
+    
+    xvfb-run -a  ./revosim \--auto 1000 
 
-You can then use e.g. SFTP to collect output files as required. 
[…]
