### 2023-07-19 b2639cc
Sort out an dxml log file save error
