### 2023-07-19 #5284 @Bisaloo
Our usual practice is to create a new release after JOSS review. In this case, I agree a patch release makes sense.

Please let me follow up with @bramvandijk88 by email to see if they are happy with the latest changes and then, you can proceed.
