+++ revosim/ui_reseed.h
@@ -1,0 +1,248 @@
+/********************************************************************************
+** Form generated from reading UI file 'reseed.ui'
+**
+** Created by: Qt User Interface Compiler version 5.15.3
+**
+** WARNING! All changes made in this file will be lost when recompiling UI file!
+********************************************************************************/
+
+#ifndef UI_RESEED_H
+#define UI_RESEED_H
+
+#include <QtCore/QVariant>
+#include <QtWidgets/QApplication>
+#include <QtWidgets/QCheckBox>
+#include <QtWidgets/QDialog>
+#include <QtWidgets/QDialogButtonBox>
+#include <QtWidgets/QHBoxLayout>
+#include <QtWidgets/QLabel>
+#include <QtWidgets/QPlainTextEdit>
+#include <QtWidgets/QRadioButton>
+#include <QtWidgets/QSpacerItem>
+#include <QtWidgets/QVBoxLayout>
+
+QT_BEGIN_NAMESPACE
+
+class Ui_Reseed
+{
+public:
+    QVBoxLayout *verticalLayout;
+    QLabel *label_3;
+    QLabel *label_4;
+    QHBoxLayout *dualReseedLayout;
+    QCheckBox *checkBoxDualReseed;
+    QRadioButton *identicalRadio;
+    QRadioButton *randomRadio;
+    QSpacerItem *verticalSpacer;
+    QLabel *label_8;
+    QCheckBox *checkBoxGenetic;
+    QLabel *label_9;
+    QHBoxLayout *tieringRadioLayout;
+    QRadioButton *radio1Tier;
+    QRadioButton *radio3Tier;
+    QRadioButton *radio5Tier;
+    QLabel *label_5;
+    QLabel *label_6;
+    QCheckBox *CheckBoxReseedSession;
+    QLabel *label_7;
+    QHBoxLayout *hexRadioLayout;
+    QRadioButton *hexRadio;
+    QRadioButton *binaryRadio;
+    QLabel *label;
+    QPlainTextEdit *genomeTextEdit;
+    QLabel *label_2;
+    QVBoxLayout *genomesLayout;
+    QDialogButtonBox *buttonBox;
+
+    void setupUi(QDialog *Reseed)
+    {
+        if (Reseed->objectName().isEmpty())
+            Reseed->setObjectName(QString::fromUtf8("Reseed"));
+        Reseed->resize(717, 788);
+        verticalLayout = new QVBoxLayout(Reseed);
+        verticalLayout->setObjectName(QString::fromUtf8("verticalLayout"));
+        label_3 = new QLabel(Reseed);
+        label_3->setObjectName(QString::fromUtf8("label_3"));
+        QFont font;
+        font.setBold(true);
+        font.setWeight(75);
+        label_3->setFont(font);
+
+        verticalLayout->addWidget(label_3);
+
+        label_4 = new QLabel(Reseed);
+        label_4->setObjectName(QString::fromUtf8("label_4"));
+        label_4->setWordWrap(true);
+
+        verticalLayout->addWidget(label_4);
+
+        dualReseedLayout = new QHBoxLayout();
+        dualReseedLayout->setObjectName(QString::fromUtf8("dualReseedLayout"));
+        checkBoxDualReseed = new QCheckBox(Reseed);
+        checkBoxDualReseed->setObjectName(QString::fromUtf8("checkBoxDualReseed"));
+
+        dualReseedLayout->addWidget(checkBoxDualReseed);
+
+        identicalRadio = new QRadioButton(Reseed);
+        identicalRadio->setObjectName(QString::fromUtf8("identicalRadio"));
+
+        dualReseedLayout->addWidget(identicalRadio);
+
+        randomRadio = new QRadioButton(Reseed);
+        randomRadio->setObjectName(QString::fromUtf8("randomRadio"));
+
[…]

[…]
