### 2023-07-22 37f4ee3
Added command line option maxthreads [int]
