### 2023-07-24 3b5ee5d
Update docs, get rid of an annoying extraneous report to command line on launch
