+++ revosim/docs_build/_sources/commandline.rst.txt
@@ -11,7 +11,7 @@ Command line switches are documented below, and examples are given of their usag
 
 Note in particular the --auto command, which automatically starts a simulation, and exits the program at its end. This will be required in most batch scenarios, as will the -k switch to turn on one or both types of logging.
 
-e.g. a command line to run a similation on the environment files in "c:\\revosim\\data\\env1" for 50000 iterations, using normal (running) logs, and with settings at defaults except for 'toroidal' (on):
+e.g. a command line to run a simulation on the environment files in "c:\\revosim\\data\\env1" for 50000 iterations, using normal (running) logs, and with settings at defaults except for 'toroidal' (on):
 
 .. code-block:: console
     
@@ -69,12 +69,12 @@ Full list of single-letter switches (also available using switch -h, help):
 - -c, --breedcost <cost (integer)>                        Breed cost.
 - -d, --maxdifftobreed <maxdifftobreed (integer)>         Maximum difference to breed.
 - -e, --environment <directory>                           Directory containing environment images.
-- -f, --usemaxdifftobreed <On/Off>                        Use maximum difference to breed criterion
-- -g, --breedwithinspecies <On/Off>                       Only allow breeding within a species
+- -f, --usemaxdifftobreed <On/Off>                        Use maximum difference to breed criterion.
+- -g, --breedwithinspecies <On/Off>                       Only allow breeding within a species.
 - -i, --disperal <distance (integer)>                     maximum dispersal distance.
 - -j, --outputpath <path>                                 path for output logs.
 - -k, --logtype <Phylogeny/Normal/Both>                   logs to generate (phylogeny is end run log, normal is running log).
-- -l, --excludenodescendents <On/Off>                     Exclude species without descendents from phylogeny logs
+- -l, --excludenodescendents <On/Off>                     Exclude species without descendents from phylogeny logs.
 - -m, --environmentmode <mode (Static|Once|Loop|Bounce)>  Environment file cycling mode.
 - -n, --energy <energy (integer)>                         Energy input.
 - -o, --tolerance <tolerance (integer)>                   Settle tolerance.
@@ -82,7 +82,7 @@ Full list of single-letter switches (also available using switch -h, help):
 - -q, --recalcfitness <On/Off>                            recalculate fitness each iteration.
 - -r, --refreshrate <rate (integer)>                      environment refresh rate.
 - -s, --slots <slots (integer)>                           Slots per pixel.
-- -t, --toroidal <On/Off>                                 Toroidal environment
+- -t, --toroidal <On/Off>                                 Toroidal environment.
 - -u, --mutation <chance (integer)>                       Chance of mutation (0-255).
 - -v, --csv <On/Off>                                      Use CSV format for normal log.
 - -w, --interpolate <On/Off>                              Interpolate environmental images
@@ -95,14 +95,14 @@ Long option only switches
 We ran out of letters! These require the long format, with \--. Otherwise they work as above.
 
 - --polling <rate [integer]>                              Set polling rate for logging and screen refresh.
-- --auto <iterations [integer]>                           Automatically start simulation and exit program after completion of specified number of iterations
+- --auto <iterations [integer]>                           Automatically start simulation and exit program after completion of specified number of iterations.
 - --nonspatial <On/Off>                                   Use non-spatial simulation mode.
-- --minspeciessize <size [integer]>                       Minimum species size to appear in logs
-- --fitnesstarget <target [integer]>                      Fitness Target
-- --breed <Obligate/Facultative/Variable/Asexual>         Breeding mode
-- --variablemutate <On/Off>                               Variable mutation rates
-- --nogui <On/Off>                                        Don't update GUI
-- --pathogens <On/Off>                                    Turn pathogens on or off
+- --minspeciessize <size [integer]>                       Minimum species size to appear in logs.
[…]
