### 2023-07-24 fd289dc
Don't show splash if running on auto from command - this tends to remove focus from whatever else you are doing at the same time
