### 2023-08-09 e3e7de5
Create CITATION.cff
