### 2023-08-09 9dc1fd7
Remove duplicate citation file
