+++ CITATION.cff
@@ -1,41 +1,0 @@
-cff-version: 1.2.0
-message: "If you use this software, please cite the paper below."
-authors:
-- family-names: "Garwood"
-  given-names: "Russell J."
-  orcid: "https://orcid.org/0000-0002-2803-9471"
-- family-names: "Spencer"
-  given-names: "Alan R.T."
-  orcid: "https://orcid.org/0000-0001-6590-405X"
-- family-names: "Sutton"
-  given-names: "Mark D."
-  orcid: "https://orcid.org/0000-0002-7137-7572"
-- family-names: "Furness"
-  given-names: "Euan N."
-  orcid: "https://orcid.org/0000-0001-7917-2304"
-title: "REvoSim and EnviroGen"
-version: 3.0.1
-doi: 10.5281/zenodo.2564192
-date-released: 2023-02-22
-url: "https://github.com/palaeoware/revosim"
-preferred-citation:
-  type: article
-  authors:
-  - family-names: "Garwood"
-    given-names: "Russell J."
-    orcid: "https://orcid.org/0000-0002-2803-9471"
-  - family-names: "Spencer"
-    given-names: "Alan R.T."
-    orcid: "https://orcid.org/0000-0001-6590-405X"
-  - family-names: "Sutton"
-    given-names: "Mark D."
-    orcid: "https://orcid.org/0000-0002-7137-7572"
-  doi: "10.1111/pala.12420"
-  journal: "Palaeontology"
-  month: 2
-  start: 339
-  end: 355 
-  title: "REvoSim: Organism-level simulation of macro- and microevolution"
-  issue: 3
-  volume: 62
-  year: 2019
