### 2023-08-09 #5284 @RussellGarwood
Thanks @Bisaloo - sorry for a slight delay on this, I was travelling. Checklist as follows:

* [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. - v3.0.1 [https://github.com/palaeoware/revosim/releases/tag/v3.0.1](https://github.com/palaeoware/revosim/releases/tag/v3.0.1)
* [x]  Archive the release on Zenodo/figshare/etc and post the DOI here. doi: 10.5281/zenodo.8228938

Outstanding point reflects the fact the paper needs to be updated to reflect the patch release which @Euan-Furness is working on. I will post again when this is complete.
