### 2023-08-09 abb3f9a
Update paper.md
Updated doi and github links from REvoSim v2 to v3.0.1. Changed paper text referrals and title from v3.0.0 to v3.0.1. Checked author ORCIDs and titles
