### 2023-08-09 #5284 @Euan-Furness
Thanks @RussellGarwood and @Bisaloo. That fourth checkbox is complete now too: all references are now to v3.0.1 throughout, and the links in the paper now link to the v3.0.1 release.
