### 2023-08-20 8a73d6d
Start generate for batch
-- Move some headers around so they are in cpp files
-- Add run for button and connect slot
