### 2023-08-20 d690781
Batch mode works
-- Run for batch now works
-- Also renamed a lot of structures, classes and variables for
consistency
