+++ utility/envirogen/mainwindow.cpp
@@ -34,6 +34,7 @@
 #include <QStandardPaths>
 #include <QShortcut>
 #include <QDebug>
+#include <QInputDialog>
 
 MainWindow *MainWin;
 
@@ -50,24 +51,23 @@ MainWindow::MainWindow(QWidget *parent) :
 
     //RJG - Globals for simulation
     generations = 500;
-    currentGeneration = 0;
+    iterations = 0;
+    runs = 0;
     stackOneSize = 0;
     stackTwoSize = 0;
     stop_flag = false;
     pause_flag = false;
-    Directory.setPath(QStandardPaths::writableLocation(QStandardPaths::DesktopLocation));
-    ui->path->setText(Directory.path());
+    ui->path->setText(QStandardPaths::writableLocation(QStandardPaths::DesktopLocation));
 
     //RJG - set up graphics
     env_image = new QImage(100, 100, QImage::Format_RGB32);
     env_image->fill(QColor(42, 42, 42));
     env_item = new QGraphicsPixmapItem(QPixmap::fromImage(*env_image));
 
-    envscene = new EnvironmentScene;
-    envscene->mw = this;
-    envscene->addItem(env_item);
-
-    ui->enviroView->setScene(envscene);
+    environmentScene = new EnvironmentScene;
+    environmentScene->mw = this;
+    environmentScene->addItem(env_item);
+    ui->enviroView->setScene(environmentScene);
 
     //RJG - Populate labels in GUI as required
     ui->envSettings->setWindowTitle("Environmental Settings");
@@ -115,12 +115,12 @@ MainWindow::MainWindow(QWidget *parent) :
     startButton = new QAction(QIcon(QPixmap(":/darkstyle/icon_play_button_green.png")), QString("Run"), this);
     startButton->setToolTip(tr("<font>Use this button to generate an environment.</font>"));
     ui->toolBar->addAction(startButton);
-    QObject::connect(startButton, SIGNAL(triggered()), this, SLOT(generateEnvironment()));
+    QObject::connect(startButton, SIGNAL(triggered()), this, SLOT(runPressed()));
 
     runForBatchButton = new QAction(QIcon(QPixmap(":/darkstyle/icon_play_batch_button_green.png")), QString("Run batch"), this);
     runForBatchButton->setToolTip(tr("<font>Use this button to generate multiple replicates of an environment.</font>"));
     ui->toolBar->addAction(runForBatchButton);
-    QObject::connect(runForBatchButton, SIGNAL (triggered()), this, SLOT (generateEnvironmentBatch()));
+    QObject::connect(runForBatchButton, SIGNAL (triggered()), this, SLOT (runBatchPressed()));
 
     stopButton = new QAction(QIcon(QPixmap(":/darkstyle/icon_stop_button_red.png")), QString("Stop"), this);
     stopButton->setToolTip(tr("<font>Use this button stop the environmental generation.</font>"));
@@ -149,7 +149,7 @@ MainWindow::MainWindow(QWidget *parent) :
     QObject::connect(aboutButton, SIGNAL (triggered()), this, SLOT (about()));
 
     //RJG - Other signal/slot connections
-    QObject::connect(ui->change_path, SIGNAL (clicked()), this, SLOT (change_path()));
+    QObject::connect(ui->change_path, SIGNAL (clicked()), this, SLOT (changePath()));
     QObject::connect(ui->settings_tab_widget, SIGNAL (currentChanged(int)), this, SLOT (tab_changed(int)));
     //RJG - these ones are pretty simple. Use lamdas.
     connect(ui->noiseMin, (void(QSpinBox::*)(int))&QSpinBox::valueChanged, this, [ = ](const int &i)
[…]
