+++ utility/envirogen/mainwindow.cpp
@@ -55,8 +55,8 @@ MainWindow::MainWindow(QWidget *parent) :
     runs = 0;
     stackOneSize = 0;
     stackTwoSize = 0;
-    stop_flag = false;
-    pause_flag = false;
+    stopFlag = false;
+    pauseFlag = false;
     ui->path->setText(QStandardPaths::writableLocation(QStandardPaths::DesktopLocation));
 
     //RJG - set up graphics
@@ -184,7 +184,7 @@ void MainWindow::tab_changed(int index)
 
 void MainWindow::runPressed()
 {
-    QString path = setupSaveDirectory();
+    QString path = setupSaveDirectory(runs);
     if (path.length() < 2) return;
     generateEnvironment(ui->environment_comboBox->currentIndex(), path);
     runs++;
@@ -194,17 +194,17 @@ void MainWindow::runBatchPressed()
 {
     bool ok;
     int runBatchFor = QInputDialog::getInt(this, "Batch...", "How many runs?", 25, 1, 999, 1, &ok);
-    if (!ok)return;
+    if (!ok) return;
 
     for (runs = 0; runs < runBatchFor; runs++)
     {
-        QString path = setupSaveDirectory();
+        QString path = setupSaveDirectory(runs);
         if (path.length() < 2) return;
         generateEnvironment(ui->environment_comboBox->currentIndex(), path);
     }
 }
 
-QString MainWindow::setupSaveDirectory()
+QString MainWindow::setupSaveDirectory(int runsLocal)
 {
 
     QDir saveDirectory(ui->path->toPlainText());
@@ -219,7 +219,7 @@ QString MainWindow::setupSaveDirectory()
 
     //RJG - Set up save directory
 
-    QString folderString(QString(PRODUCTNAME) + "_output_" + QString("%1").arg(runs, 3, 10, QChar('0')) + QDir::separator());
+    QString folderString(QString(PRODUCTNAME) + "_output_" + QString("%1").arg(runsLocal, 3, 10, QChar('0')) + QDir::separator());
     if (!saveDirectory.mkpath(folderString))
     {
         QMessageBox::warning(this, "Error", "Cant save images. Permissions issue?");
@@ -231,60 +231,59 @@ QString MainWindow::setupSaveDirectory()
 }
 
 //RJG - Generates environment based on which tab is selected in the tab dock widget.
-void MainWindow::generateEnvironment(int environmentType, QString path)
+void MainWindow::generateEnvironment(int environmentType, QString path, bool batch)
 {
+    //RJG - new environment object
+    EnvironmentClass *environmentObject = nullptr;
 
     //RJG - Select which kind of environment object to create - all inheret environmentclass, which has the random number and save functions in it
-    if (environmentType == 0) // Russell environment
-        environmentObject = new russellenvironment;
-
-    if (environmentType == 1) //Mark environment
-        environmentObject = new markenvironment;
-
-    if (environmentType == 2)   //Noise stack
+    switch (environmentType)
     {
+    case 0:
+        environmentObject = new russellenvironment; // Russell environment
+        break;
+    case 1:
+        environmentObject = new markenvironment; //Mark environment
+        break;
+    case 2: //Noise stack
         environmentObject = new noiseenvironment;
         if (MainWin->ui->noiseMin->value() >= MainWin->ui->noiseMax->value())
         {
-            QMessageBox::warning(nullptr, "Error", "Min is greater than Max - please change this before proceeding.", QMessageBox::Ok);
-            reset_gui();
+            QMessageBox::warning(this, "Error", "Min is greater than Max - please change this before proceeding.", QMessageBox::Ok);
+            reset(environmentObject);
             return;
         }
-    }
-
[…]
