+++ utility/envirogen/mainwindow.cpp
@@ -35,6 +35,7 @@
 #include <QShortcut>
 #include <QDebug>
 #include <QInputDialog>
+#include <QtConcurrent>
 
 MainWindow *MainWin;
 
@@ -163,8 +164,15 @@ MainWindow::MainWindow(QWidget *parent) :
     ui->noiseMin->setMaximum(ui->noiseMax->value() - 1);
     ui->noiseMax->setMinimum(ui->noiseMin->value() + 1);
 
+    saveImages = true;
+    ui->save_images_checkbox->setChecked(saveImages);
+    connect(ui->save_images_checkbox, &QAbstractButton::toggled, this, [ = ](const bool & i)
+    {
+        saveImages = i;
+    });
+
     //RJG - shortcuts
-    new QShortcut(QKeySequence(Qt::CTRL + Qt::Key_R), this, SLOT(generateEnvironment()));
+    new QShortcut(QKeySequence(Qt::CTRL + Qt::Key_R), this, SLOT(runPressed()));
     new QShortcut(QKeySequence(Qt::CTRL + Qt::Key_P), this, SLOT(pause()));
     new QShortcut(QKeySequence(Qt::Key_Escape), this, SLOT(stop()));
 }
@@ -186,22 +194,65 @@ void MainWindow::runPressed()
 {
     QString path = setupSaveDirectory(runs);
     if (path.length() < 2) return;
-    generateEnvironment(ui->environment_comboBox->currentIndex(), path);
+    generateEnvironment(ui->environment_comboBox->currentIndex(), path, ui->spinSize->value(), ui->spinSize->value());
     runs++;
 }
 
 void MainWindow::runBatchPressed()
 {
+
+    if (ui->environment_comboBox->currentIndex() > 2)
+    {
+        QMessageBox::warning(this, "Sorry!", "Batch runs are not currently available for your chosen environment type. Please request the developers add it.");
+        return;
+    }
+
     bool ok;
     int runBatchFor = QInputDialog::getInt(this, "Batch...", "How many runs?", 25, 1, 999, 1, &ok);
     if (!ok) return;
 
-    for (runs = 0; runs < runBatchFor; runs++)
+    QVector<int> runsList(runBatchFor);
+    std::iota(runsList.begin(), runsList.end(), 0);
+
+    auto count = 0;
+    bool batchRunning = true;
+    do
     {
-        QString path = setupSaveDirectory(runs);
-        if (path.length() < 2) return;
-        generateEnvironment(ui->environment_comboBox->currentIndex(), path);
+        //In a previous version RJG had used the progress bar in the status bar, but connecting the future watcher signals and slots provide very challenging
+        //Hence now do this with a dialogue - first create a dialogue, then QFutureWatcher and connect signals and slots.
+        QProgressDialog dialog;
+        dialog.setLabelText(QString("Starting a batch run of %1 simulations on %2 cores.").arg(runsList.length()).arg(QThread::idealThreadCount()));
+        QFutureWatcher<void> futureWatcher;
+        QObject::connect(&futureWatcher, &QFutureWatcher<void>::finished, &dialog, &QProgressDialog::reset);
+        QObject::connect(&dialog, &QProgressDialog::canceled, &futureWatcher, &QFutureWatcher<void>::cancel);
+        QObject::connect(&futureWatcher, &QFutureWatcher<void>::progressRangeChanged, &dialog, &QProgressDialog::setRange);
+        QObject::connect(&futureWatcher, &QFutureWatcher<void>::progressValueChanged,  &dialog, &QProgressDialog::setValue);
+
+        //Do the runs using QtConcurrent::filter which modified the sequence in place
[…]
