+++ utility/envirogen/mainwindow.cpp
@@ -244,15 +244,6 @@ void MainWindow::runBatchPressed()
     }
     //Run up to 1 times so this cannot get caught in an infinite loop
     while (runsList.count() > 0 && count < 1 && batchRunning);
-
-
-    /*  for (runs = 0; runs < runBatchFor; runs++)
-      {
-          QString path = setupSaveDirectory(runs);
-          if (path.length() < 2) return;
-          generateEnvironment(ui->environment_comboBox->currentIndex(), path, ui->spinSize->value(), ui->spinSize->value(), true);
-      }
-      */
 }
 
 QString MainWindow::setupSaveDirectory(int runsLocal)
