### 2023-08-22 4b83164
Improving parallelisation
-- EnvSettings are now stored in a structure, implemented for russell
environment
-- there is a global pointer to a settings object in the simulation
-- Settings are read from mainwindow on construct, so can be created
just before needed
-- russell environment constructor now requires settings object to run
-- Added this to generate function as well
-- next to check this actually works with environment, and implement for
mark environmnet
