+++ utility/envirogen/markenvironment.cpp
@@ -33,10 +33,10 @@ markenvironment::markenvironment()
     {
         objxpos[i] = MainWin->simulationRandoms->randDouble() * MainWin->ui->spinSize->value();
         objypos[i] = MainWin->simulationRandoms->randDouble() * MainWin->ui->spinSize->value();
-        objxvel[i] = MainWin->simulationRandoms->randDouble() * maxvel * 2 - maxvel;
-        objyvel[i] = MainWin->simulationRandoms->randDouble() * maxvel * 2 - maxvel;
-        objsize[i] = MainWin->simulationRandoms->randDouble() * (maxsize - minsize) + minsize;
-        objsizevel[i] = MainWin->simulationRandoms->randDouble() * maxsizevel * 2 - maxsizevel;
+        objxvel[i] = MainWin->simulationRandoms->randDouble() * maximumVelocity * 2 - maximumVelocity;
+        objyvel[i] = MainWin->simulationRandoms->randDouble() * maximumVelocity * 2 - maximumVelocity;
+        objsize[i] = MainWin->simulationRandoms->randDouble() * (maxSize - minsize) + minsize;
+        objsizeVelocity[i] = MainWin->simulationRandoms->randDouble() * maxsizeVelocity * 2 - maxsizeVelocity;
         for (int o = 0; o < 3; o++)
             objcolours[i][o] = (double)((int)MainWin->simulationRandoms->rand8() - 127);
 
@@ -54,10 +54,10 @@ markenvironment::markenvironment()
 void markenvironment::ReadSettings()
 {
     objectcount = MainWin->ui->ME_Objects->value();
-    maxsize = (double)(MainWin->ui->ME_MaxSize->value());
+    maxSize = (double)(MainWin->ui->ME_maxSize->value());
     minsize = (double)(MainWin->ui->ME_MinSize->value());
-    maxvel = MainWin->ui->ME_MaxVel->value();
-    maxsizevel = MainWin->ui->ME_MaxSizeVel->value();
+    maximumVelocity = MainWin->ui->ME_MaxVel->value();
+    maxsizeVelocity = MainWin->ui->ME_maxSizeVel->value();
     maxcolvel = MainWin->ui->ME_MaxColVel->value();
     maxtightvel = MainWin->ui->ME_MaxTightVel->value();
     maxtight = MainWin->ui->ME_MaxTight->value();
@@ -66,7 +66,7 @@ void markenvironment::ReadSettings()
     tighttweak = MainWin->ui->ME_MaxTightAccel->value();
     coltweak = MainWin->ui->ME_MaxColAccel->value();
     veltweak = MainWin->ui->ME_MaxAccel->value();
-    sizetweak = MainWin->ui->ME_MaxSizeAccel->value();
+    sizetweak = MainWin->ui->ME_maxSizeAccel->value();
     iter_reset = MainWin->ui->ME_Iterations->value();
 
 }
@@ -115,16 +115,16 @@ void markenvironment::regenerate()
         }
 
         //add size change, apply limits
-        objsize[i] += (objsizevel[i] * speedfactor);
-        if (objsize[i] > maxsize)
+        objsize[i] += (objsizeVelocity[i] * speedfactor);
+        if (objsize[i] > maxSize)
         {
-            objsize[i] = maxsize;
-            objsizevel[i] = 0;
+            objsize[i] = maxSize;
+            objsizeVelocity[i] = 0;
         }
         if (objsize[i] < minsize)
         {
             objsize[i] = minsize;
-            objsizevel[i] = 0;
+            objsizeVelocity[i] = 0;
         }
 
         //do colour
[…]
