### 2023-08-22 b606fc9
Env settings implementation
-- Fixed error in russell environment, all works.
-- Made it so run works with the new pointer approach
-- Note this will have broken all other environments for the moment
