### 2023-08-22 a79e313
Remove extraneous code from environment class
Not sure why this is in there
