### 2023-08-22 b46e61a
Sort out all environment classes to accept settings in constructor
Fix some inheritance issues associated with this
Change calls from mainwindow.cpp
Next up, add X Y and path to object, move save image to environment
class, and then implement the other environments
