+++ utility/envirogen/colour.h
@@ -20,10 +20,12 @@
 
 #include <QColor>
 
+#include "environmentclass.h"
+
 class colour : public EnvironmentClass
 {
 public:
-    colour();
+    colour(EnvironmentSettings constructorSettings);
     void regenerate();
 
 private:
