### 2023-08-23 3824960
Continuing to OO things
-- X and Y now belong to environment object, as does save path - needs
implementing with a save function, and possibly setupsavedirectory
-- Need to shift randoms from mainwindow through to each environment
object
