### 2023-08-24 034fffa
Moved randoms through to environmentclass.cpp
These no longer belong to main window
