### 2023-08-24 7554033
Implement mark environment
-- Save settings from GUI into settings object
-- Define settings in constrcutor
-- Replace all refs to mainwindow

Next up sort up call from mainwindow
