### 2023-08-25 8aeb2b6
Replace colourmap with a vector
this is horrendously slow, presumably because of all the allocation -
assume it'll be much better if allocated in constructor
