### 2023-08-26 5d6da70
Try and remove requirement for global mainwin, as well as finish implementing other environments
Currently won't compile due to a few missing mainwin refs
