+++ utility/envirogen/russellenvironment.cpp
@@ -54,6 +54,29 @@ russellenvironment::russellenvironment(EnvironmentSettings constructorSettings) 
         seeds[i].size = ((double)r * (double)maxSize) / 256.;
         seeds[i].initialised = true;
     }
+
+    //Allocate structures
+    for (int i = 0; i < x; i++)
+    {
+        colourMap.append(QVector<QVector<double>>());
+        templaplace.append(QVector<QVector<bool>>());
+        laplace.append(QVector<int>());
+        for (int j = 0; j < y; j++)
+        {
+            colourMap[i].append(QVector <double>());
+            templaplace[i].append(QVector<bool>());
+            laplace[i].append(0);
+            for (int k = 0; k < 3; k++)
+            {
+                colourMap[i][j].append(static_cast<double>(environment[i][j][k]));
+            }
+
+            for (int k = 0; k < nSeed; k++)
+            {
+                templaplace[i][j].append(false);
+            }
+        }
+    }
 }
 
 
@@ -123,77 +146,52 @@ void russellenvironment::regenerate()
     }
 
     //Interpolation then occurs in the laplace function
-    laplace();
+    doLaplace();
 }
 
-void russellenvironment::laplace()
+void russellenvironment::doLaplace()
 {
 
     //Interpolate
-    //First fill colours
-
-    //TODO - fix all multidimensional arrays to what we see in
-    //https://stackoverflow.com/questions/1946830/multidimensional-variable-size-array-in-c
-
-
-    QVector<QVector<QVector<double>>> colourMap;
-
-    for (int i = 0; i < x; i++)
-    {
-        colourMap.append(QVector<QVector<double>>());
-        for (int j = 0; j < y; j++)
-        {
-            colourMap[i].append(QVector <double>());
-            for (int k = 0; k < 3; k++)
-            {
-                colourMap[i][j].append(static_cast<double>(environment[i][j][k]));
-            }
-        }
-    }
-
-    //Do it all in double colourMap so don't get errors from using environment (integers)
-    int laplace[MainWin->ui->spinSize->value()][MainWin->ui->spinSize->value()];
     double eTotal, e[3];
     //Laplacian = residual, total and then residual for R,G and B
 
+    //First fill colours
     //Initialise colourmap from environment to make laplacian faster
     for (int n = 0; n < x; n++)
         for (int m = 0; m < y; m++)
         {
             laplace[n][m] = 0;
             for (int i = 0; i < 3; i++)colourMap[n][m][i] = environment[n][m][i];
+            for (int i = 0; i < nSeed; i++)templaplace[n][m][i] = false;
         }
 
-    double x, y;
+    double localX, localY;
+
     for (int l = 0; l < nSeed; l++)
     {
-        bool templaplace[MainWin->ui->spinSize->value()][MainWin->ui->spinSize->value()];
-        for (int n = 0; n < MainWin->ui->spinSize->value(); n++)
-            for (int m = 0; m < MainWin->ui->spinSize->value(); m++)
-                templaplace[n][m] = false;
-
         for (double z = -PI; z < PI; z += .01)
         {
             //Draw circles...
-            x = seeds[l].n + (seeds[l].size * cos(z));
-            y = seeds[l].m + (seeds[l].size * sin(z));
+            localX = seeds[l].n + (seeds[l].size * cos(z));
+            localY = seeds[l].m + (seeds[l].size * sin(z));
 
             if (periodic)
             {
-                if (y > (MainWin->ui->spinSize->value() - 1)) y -= (MainWin->ui->spinSize->value() - 1);
[…]
