### 2023-08-26 50bf112
Fix all pointers that use mainWin global to local
