+++ utility/envirogen/about.h
@@ -29,7 +29,7 @@ class About : public QDialog
     Q_OBJECT
 
 public:
-    explicit About(QWidget *parent = 0);
+    explicit About(QWidget *parent);
     ~About();
 
 private:
