### 2023-08-26 b9d08f8
Fiddline with main window
