### 2023-08-26 c9c679f
Cleaning up from rewrite
All done now I think
