+++ utility/envirogen/mainwindow.cpp
@@ -37,6 +37,16 @@
 #include <QInputDialog>
 #include <QtConcurrent>
 
+/*********** TREvoSim structure - overview ***********/
+/* Following a rewrite by RJG in August 2023, the classes interacting in Envirogen are as follows:
+ * -- Environment class - this is a base class from which the different environments derive. It stores some of the settings e.g. X and Y coords
+ * -- Each type of environment inherits this, and then adds its own settings, and has a rewritten regenerate function
+ * -- In order to allow batches to be run in parallel, these have been largely divorced from the GUI - unless they are not running in batch mode
+ * -- To this end there is a class environment settings, which at the point generation is initiated, places all the of the readings from the GUI into an object of this class
+ * -- This is then passed to the constructor of each derived environment object (as well as the base class), allowing each replicate to run independently of the GUI
+ * The only tab that relies on the GUI now is combine stacks. The rest are either available as batch mode or could be very easily if we so wish.
+ */
+/****************************************************/
 MainWindow::MainWindow(QWidget *parent) :
     QMainWindow(parent),
     ui(new Ui::MainWindow)
@@ -189,11 +199,10 @@ void MainWindow::tab_changed(int index)
 
 void MainWindow::runPressed()
 {
-    QString path = setupSaveDirectory(runs);
-    if (path.length() < 2) return;
     currentEnvironmentSettings = new EnvironmentSettings(this);
-    currentEnvironmentSettings->savePath = path; //Eventually we can do this above, but for compatability between environments, currently do both
-    generateEnvironment(ui->environment_comboBox->currentIndex(), path, ui->spinSize->value(), ui->spinSize->value(), *currentEnvironmentSettings);
+    currentEnvironmentSettings->savePath =  setupSaveDirectory(runs);
+    if (currentEnvironmentSettings->savePath.length() < 2)return;
+    generateEnvironment(ui->environment_comboBox->currentIndex(), *currentEnvironmentSettings);
     delete currentEnvironmentSettings;
     runs++;
 }
@@ -234,11 +243,10 @@ void MainWindow::runBatchPressed()
         //Do the runs using QtConcurrent::filter which modified the sequence in place
         futureWatcher.setFuture(QtConcurrent::filter(runsList, [this](const int &run)
         {
-            QString path = setupSaveDirectory(run);
-            if (path.length() < 2) return false;
             EnvironmentSettings localEnvironmentSettings = *currentEnvironmentSettings;
-            localEnvironmentSettings.savePath = path; //Eventually we can do this above, but for compatability between environments, currently do both
-            return generateEnvironment(ui->environment_comboBox->currentIndex(), path, ui->spinSize->value(), ui->spinSize->value(), localEnvironmentSettings, true);
+            localEnvironmentSettings.savePath =  setupSaveDirectory(run);
+            if (localEnvironmentSettings.savePath.length() < 2)return false;
+            return generateEnvironment(ui->environment_comboBox->currentIndex(), localEnvironmentSettings, true);
         }));
 
         // Display the dialog and start the event loop.
@@ -250,19 +258,10 @@ void MainWindow::runBatchPressed()
     //Run up to 1 times so this cannot get caught in an infinite loop
     while (runsList.count() > 0 && count < 1 && batchRunning);
     delete currentEnvironmentSettings;
-
-    /*  for (runs = 0; runs < runBatchFor; runs++)
-      {
-          QString path = setupSaveDirectory(runs);
-          if (path.length() < 2) return;
[…]
