+++ utility/envirogen/build/moc_predefs.h
@@ -1,403 +1,0 @@
-#define __SSP_STRONG__ 3
-#define __DBL_MIN_EXP__ (-1021)
-#define __cpp_attributes 200809L
-#define __UINT_LEAST16_MAX__ 0xffff
-#define __ATOMIC_ACQUIRE 2
-#define __FLT128_MAX_10_EXP__ 4932
-#define __FLT_MIN__ 1.17549435082228750796873653722224568e-38F
-#define __GCC_IEC_559_COMPLEX 2
-#define __UINT_LEAST8_TYPE__ unsigned char
-#define __SIZEOF_FLOAT80__ 16
-#define __INTMAX_C(c) c ## L
-#define __CHAR_BIT__ 8
-#define __UINT8_MAX__ 0xff
-#define __SCHAR_WIDTH__ 8
-#define __WINT_MAX__ 0xffffffffU
-#define __FLT32_MIN_EXP__ (-125)
-#define __cpp_static_assert 200410L
-#define __ORDER_LITTLE_ENDIAN__ 1234
-#define __SIZE_MAX__ 0xffffffffffffffffUL
-#define __WCHAR_MAX__ 0x7fffffff
-#define __GCC_HAVE_SYNC_COMPARE_AND_SWAP_1 1
-#define __GCC_HAVE_SYNC_COMPARE_AND_SWAP_2 1
-#define __GCC_HAVE_SYNC_COMPARE_AND_SWAP_4 1
-#define __DBL_DENORM_MIN__ double(4.94065645841246544176568792868221372e-324L)
-#define __GCC_HAVE_SYNC_COMPARE_AND_SWAP_8 1
-#define __GCC_ATOMIC_CHAR_LOCK_FREE 2
-#define __GCC_IEC_559 2
-#define __FLT32X_DECIMAL_DIG__ 17
-#define __FLT_EVAL_METHOD__ 0
-#define __cpp_binary_literals 201304L
-#define __FLT64_DECIMAL_DIG__ 17
-#define __CET__ 3
-#define __GCC_ATOMIC_CHAR32_T_LOCK_FREE 2
-#define __cpp_variadic_templates 200704L
-#define __UINT_FAST64_MAX__ 0xffffffffffffffffUL
-#define __SIG_ATOMIC_TYPE__ int
-#define __DBL_MIN_10_EXP__ (-307)
-#define __FINITE_MATH_ONLY__ 0
-#define __FLT32X_MAX_EXP__ 1024
-#define __FLT32_HAS_DENORM__ 1
-#define __UINT_FAST8_MAX__ 0xff
-#define __cpp_rvalue_reference 200610L
-#define __FLT32_MAX_10_EXP__ 38
-#define __DEC64_MAX_EXP__ 385
-#define __INT8_C(c) c
-#define __INT_LEAST8_WIDTH__ 8
-#define __UINT_LEAST64_MAX__ 0xffffffffffffffffUL
-#define __INT_LEAST8_MAX__ 0x7f
-#define __SHRT_MAX__ 0x7fff
-#define __LDBL_MAX__ 1.18973149535723176502126385303097021e+4932L
-#define __FLT64X_MAX_10_EXP__ 4932
-#define __LDBL_IS_IEC_60559__ 2
-#define __FLT64X_HAS_QUIET_NAN__ 1
-#define __UINT_LEAST8_MAX__ 0xff
-#define __GCC_ATOMIC_BOOL_LOCK_FREE 2
-#define __FLT128_DENORM_MIN__ 6.47517511943802511092443895822764655e-4966F128
-#define __UINTMAX_TYPE__ long unsigned int
-#define __linux 1
[…]
