+++ utility/envirogen/docs_build/_sources/introduction.rst.txt
@@ -30,9 +30,11 @@ An image stack can be generated by hitting run on the top toolbar. This is store
 .. figure:: _static/toolbar.png
     :align: center
 
+
 The toolbar has the following options:
 
-:Run: Generate environemtnal images.
+:Run: Generate environmental images.
+:Run batch: Create multiple replicates of environmental images in parallel.
 :Stop: Stop generating images.
 :Pause: Temporarily stop generating images.
 :About: Launch dialogue with information about EnviroGen.

[…]
