+++ JOSS/paper.md
@@ -80,7 +80,7 @@ The species identification algorithm has been recoded to improve performance.
 
 # Statement of need
 
-Despite simulations becoming progressively more common in evolutionary ecology [@zurrell:2010; @cabral:2017; @barido-sottani:2020], many function at the level of the population, rather than the individual [@rangel:2018; @hagen:2021; @reddin:2022]. Simulations that do function at the level of the individual often focus on ecology, rather than evolution [@smith:2010; @harfoot:2014], although other individual-based evolutionary simulation tools do exist [@dolson:2021]. All digital evolution and simulation tools experience a tradeoff between complexity and efficiency [@dolson:2021], and REvoSim is no exception. REvoSim minimises system complexity through abstraction of organism and envrionment traits: an approach that improves efficiency, while still allowing for the investigation of complex phenomena as emergent properties of the model. This approach allows REvoSim to simulate long time spans and large population sizes on modest hardware. However, this approach abstracts out any effect of genetic complexity or ontogeny on organisms. REvoSim is therefore complementary with other simulation and digital evolution tools that incorporate greater explicit complexity (particularly genetic complexity; [@batut:2013]) at the cost of efficiency.
+Despite simulations becoming progressively more common in evolutionary ecology [@zurrell:2010; @cabral:2017; @barido-sottani:2020], many function at the level of the population, rather than the individual [@rangel:2018; @hagen:2021; @reddin:2022]. Simulations that do function at the level of the individual often focus on ecology, rather than evolution [@smith:2010; @harfoot:2014], although other individual-based evolutionary simulation tools do exist [@dolson:2021]. All digital evolution and simulation tools experience a tradeoff between complexity and efficiency [@dolson:2021], and REvoSim is no exception. REvoSim minimises system complexity through abstraction of organism and environment traits: an approach that improves efficiency, while still allowing for the investigation of complex phenomena as emergent properties of the model. This approach allows REvoSim to simulate long time spans and large population sizes on modest hardware. However, this approach abstracts out any effect of genetic complexity or ontogeny on organisms. REvoSim is therefore complementary with other simulation and digital evolution tools that incorporate greater explicit complexity (particularly genetic complexity; [@batut:2013]) at the cost of efficiency.
 
 The new features added in REvoSim v3 allow for the study of ecological processes (predation, symbioses, and pathogens) that have been implicated in the evolution of diversity gradients, such as the latitudinal biodiversity gradient [@johnson:2012; @roslin:2017]. Studies could directly modify the intensity or operation of these processes within REvoSim in order to determine their impact on the formation of such gradients. Furthermore, these processes allow REvoSim to be used as a tool for complex ecological simulation in the absence of evolutionary processes, and have also increased the versatility of the tool throughout.
 
