### 2023-08-28 fbc906b
Remove duplicated figure numbering
