+++ JOSS/paper.md
@@ -58,15 +58,15 @@ REvoSim v3 adds mechanism that allows for the productivity and/or mutation rates
 Inter-organism interactions can be simulated in REvoSim v3 using one of two approaches, referred to as the ‘genome blocks’ and ‘XOR’ mechanisms. Both interaction mechanisms are designed such that small genome changes result in small changes to interaction scores. In both cases, the number of attempted interactions per-timestep per-individual is configurable. Individuals attempt to interact by searching for another individual at the same position as them in the environment. The probability of success of this search increases as organism-count in the cell increases. If another individual is found, an interaction occurs. Under ‘genome blocks’, an ‘interaction score’ is derived from a configurable table that describes how every possible combination of two-bit organism genome components interact \autoref{fig:figure1}. Under ‘XOR’, one of the genomes is rotationally shifted by one bit both to the left and to the right \autoref{fig:figure2}, and each shifted genome is combined with an exclusive-or function with the unshifted genome. The number of 1s in the right-shifted, XORed binary sequence subtracted from the number of 1s in the left-shifted, XORed binary sequence, creates an interaction score \autoref{fig:figure2}. 
 
 
-![Figure 1. An example interactions table for the ‘genome blocks’ mechanism. Adjacent bits in an interacting organism’s genomes are paired (top) and compared with the corresponding pair in the other interacting organism (left) to determine their contribution to the interaction score. The total interaction score is the sum of the scores generated by every pair of bits.\label{fig:figure1}](figure1.png)
+![An example interactions table for the ‘genome blocks’ mechanism. Adjacent bits in an interacting organism’s genomes are paired (top) and compared with the corresponding pair in the other interacting organism (left) to determine their contribution to the interaction score. The total interaction score is the sum of the scores generated by every pair of bits.\label{fig:figure1}](figure1.png)
 
 
-![Figure 2. A flowchart depicting the XOR mechanism for two genomes (A: blue, and B: red), abbreviated to 16 bits in length for ease of interpretation. Swapping genomes A and B would reverse the sign of the interaction score.\label{fig:figure2}](figure2.png)
+![A flowchart depicting the XOR mechanism for two genomes (A: blue, and B: red), abbreviated to 16 bits in length for ease of interpretation. Swapping genomes A and B would reverse the sign of the interaction score.\label{fig:figure2}](figure2.png)
 
 Interactions may alter either organism fitnesses, modifying the rate at which they collect energy from the environment, or allow organisms to transfer energy from one another. Under the former, the interaction score (which may be negative) is added to the fitness of the instigating organism. Under the latter, the interaction score controls the proportion of the sought organism’s current energy that is lost in the interaction \autoref{fig:figure3}. Optionally, energy transfer can be disabled where this proportion would be below some threshold. Of the energy lost by the sought organism, a configurable fraction is transferred to the instigating organism, reflecting loss of energy to metabolic inefficiency [@lindeman:1942; @anderson:2009].
 
 
-![Figure 3. The proportion of energy lost by the sought organism, when direct energy transfer is occurring, for any given mean genome word interaction score.\label{fig:figure3}](figure3.png)
+![The proportion of energy lost by the sought organism, when direct energy transfer is occurring, for any given mean genome word interaction score.\label{fig:figure3}](figure3.png)
 
 To accompany the interactions system, REvoSim v3 includes the option to seed either three or five (in addition to the previous default of one) different genomes at the start of the simulation. When three or five genomes are seeded, preset genomes are used that are known to form a single, unbranching food-chain under the XOR mechanism [@barbier:2019].
 
