+++ JOSS/paper.bib
@@ -103,7 +103,7 @@
 }
 
 @book{matthews:2021,
-	doi = {10.1017/9771108569422},
+	doi = {10.1017/9781108569422},
 	Author = {{Matthews}, T.~J. and {Triantis}, K.~A. and {Whittaker}, R.~J.},
 	Booktitle = {The species-area relationship: theory and application.~ISBN: 9781108701877},
 	Publisher = {Cambridge University Press},
