### 2023-09-05 d11a762
Update paper.md
Updated affiliations from "UK" to "United Kingdom".
