### 2023-09-05 #5284 @Kevin-Mattheus-Moerman
@Euan-Furness thanks for making that fix in the affiliations. One more point remains:

- [ ] I have just noticed that your software license (GPL-3.0 license ) actually does not match the ZENODO archive listed license (GNU Affero General Public License v3.0). Can you please edit  this?
