### 2023-09-10 #5284 @Kevin-Mattheus-Moerman
@Euan-Furness congratulations on this paper in JOSS!

Thanks for editing @Bisaloo

And a special thanks to the reviewers:  @bramvandijk88, @emilydolson
